Create HTML5 VM Console Test
Test runs against VM/Instances on RHV, OpenStack and VMware.
Tests that the HTML5 VM Console feature works.   Attempts to
login to a console on a VM/Instance on the specified provider
touch a file, and verify the files existence.

Also added to cfme/common/vm.py:

    - open_console method
    - console method

A new class VMConsole is added to abstract the console functionality
of:

    - reading the console banner
    - sending text to the console
    - reading the console bitmap

All of this while managing switching from the CFME tab to the console
tab and back.

(FIXES RHCFQE-2647)